:construction: Rework truth tables and their free functions into their own translation units
